requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.36
Pillow==8.4.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
